Update to lemminx 0.29.0
It seems that, while the function signatures haven't changed,
there were some API changes during the implementation of RelaxNG
support (likely for the better). I adapted to these changes to get the
tests passing again.

We need to use `-noverify` in tests, because lemminx adds classes into the
jing packages in order to access some package private stuff.
(this breaks jar signing angry, which `-noverify` silences).

It should be fine for the built .jar, since we remove all the signatures
during shading.

Signed-off-by: David Thompson <davthomp@redhat.com>